refactor(web): [Daily Plan] daily plan hooks to use specialized queries
- Replaced `useDailyPlan` with `useDailyPlanQuery` in `page-component.tsx` for improved data fetching.
- Updated `all-plans-modal.tsx`, `all-plans.tsx`, and other components to utilize `useEmployeeDailyPlans` and `useCreateDailyPlan` for better separation of concerns.
- Refactored various components to ensure consistent usage of new hooks, enhancing maintainability and performance.
- Adjusted logic in `task-filters.tsx` and `user-profile-plans.tsx` to align with the new data structure from the updated hooks.

This change streamlines the daily plan management process and improves the overall codebase structure.
