+++ apps/web/core/components/tasks/daily-plan/future-tasks.tsx
@@ -14,7 +14,7 @@ import { DragDropContext, Draggable, Droppable } from '@hello-pangea/dnd';
 import { useDateRange } from '@/core/hooks/daily-plans/use-date-range';
 import DailyPlanTasksTableView from './table-view';
 import { HorizontalSeparator } from '../../duplicated-components/separator';
-import { useDailyPlan } from '@/core/hooks';
+import { useDailyPlanQuery } from '@/core/hooks/daily-plans/use-daily-plan-query';
 
 export function FutureTasks({
 	profile,
@@ -29,7 +29,7 @@ export function FutureTasks({
 }) {
 	// Use employeeId from props if provided, otherwise calculate from user
 	const targetEmployeeId = employeeId ?? user?.employee?.id ?? user?.employeeId ?? '';
-	const { futurePlans } = useDailyPlan(targetEmployeeId);
+	const { futurePlans } = useDailyPlanQuery(targetEmployeeId);
 	// Use a safe default instead of direct localStorage access
 	const { date } = useDateRange('Future Tasks');
 	const view = useAtomValue(dailyPlanViewHeaderTabs);
@@ -49,7 +49,7 @@ export function FutureTasks({
 		return filteredData;
 	}, [date, futurePlans, user, filterByEmployee]);
 
-	if(!futureDailyPlanTasks) return null;
+	if (!futureDailyPlanTasks) return null;
 	return (
 		<div className="flex flex-col gap-6">
 			{futureDailyPlanTasks?.length > 0 ? (
